// packages/datadog-api-client-v2/models/LogsCompute.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LogsAggregationFunction } from './LogsAggregationFunction';
import { LogsComputeType } from './LogsComputeType';
/**
* A compute rule to compute metrics or timeseries
*/
export declare class LogsCompute {
    'aggregation': LogsAggregationFunction;
    /**
    * The time buckets' size (only used for type=timeseries) Defaults to a resolution of 150 points
    */
    'interval'?: string;
    /**
    * The metric to use
    */
    'metric'?: string;
    'type'?: LogsComputeType;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsCompute;
    static serialize(data: LogsCompute): {
        [key: string]: any;
    };
    constructor();
}