// packages/datadog-api-client-v2/models/LogsGroupBy.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LogsAggregateSort } from './LogsAggregateSort';
import { LogsGroupByHistogram } from './LogsGroupByHistogram';
import { LogsGroupByMissing } from './LogsGroupByMissing';
import { LogsGroupByTotal } from './LogsGroupByTotal';
/**
* A group by rule
*/
export declare class LogsGroupBy {
    /**
    * The name of the facet to use (required)
    */
    'facet': string;
    'histogram'?: LogsGroupByHistogram;
    /**
    * The maximum buckets to return for this group by
    */
    'limit'?: number;
    'missing'?: LogsGroupByMissing;
    'sort'?: LogsAggregateSort;
    'total'?: LogsGroupByTotal;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsGroupBy;
    static serialize(data: LogsGroupBy): {
        [key: string]: any;
    };
    constructor();
}