// packages/datadog-api-client-v2/models/LogsQueryFilter.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* The search and filter query settings
*/
export declare class LogsQueryFilter {
    /**
    * The minimum time for the requested logs, supports date math and regular timestamps
    */
    'from'?: string;
    /**
    * For customers with multiple indexes, the indexes to search. Defaults to ['*'] which means all indexes.
    */
    'indexes'?: Array<string>;
    /**
    * The search query - following the log search syntax.
    */
    'query'?: string;
    /**
    * The maximum time for the requested logs, supports date math and regular timestamps
    */
    'to'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsQueryFilter;
    static serialize(data: LogsQueryFilter): {
        [key: string]: any;
    };
    constructor();
}