// packages/datadog-api-client-v2/models/LogsQueryOptions.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Global query options that are used during the query. Note: You should only supply timezone or time offset but not both otherwise the query will fail.
*/
export declare class LogsQueryOptions {
    /**
    * The time offset (in seconds) to apply to the query.
    */
    'timeOffset'?: number;
    /**
    * The timezone can be specified both as an offset, for example: \"UTC+03:00\".
    */
    'timezone'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsQueryOptions;
    static serialize(data: LogsQueryOptions): {
        [key: string]: any;
    };
    constructor();
}